Split beyla_linux.go by concern
